requirements.txt
streamlit==1.21.0
pandas==2.0.3
wordcloud==1.9.2
matplotlib==3.7.3
azure-ai-textanalytics==5.3.0
azure-core==1.26.1
